finalApp

+++ lib/services/rent/rent_service_old/profile/profile_view.dart
@@ -1,3 +1,4 @@
+// profile_view.dart
 import 'package:flutter/material.dart';
 import 'package:r_and_e_monitor/services/rent/rent_service_old/profile/create_or_update_profile.dart';
 import 'package:r_and_e_monitor/services/rent/rent_service_old/profile/profile_list_view.dart';
@@ -8,7 +9,7 @@ import '../../../cloud/profiles/read_profile.dart';
 import '../../../cloud/employee_services/cloud_rent_service.dart';
 
 class ProfileView extends StatefulWidget {
-  const ProfileView({Key? key}) : super(key: key);
+  const ProfileView({super.key});
 
   @override
   State<ProfileView> createState() => _ProfileViewState();
@@ -30,9 +31,7 @@ class _ProfileViewState extends State<ProfileView> {
       appBar: AppBar(title: const Text("Profiles"), actions: [
         IconButton(
           onPressed: () {
-            Navigator.of(context).pushNamed(
-              createOrUpdateProfileRoute,
-            );
+            Navigator.of(context).pushNamed(createOrUpdateProfileRoute);
           },
           icon: const Icon(Icons.add),
         ),
@@ -45,35 +44,76 @@ class _ProfileViewState extends State<ProfileView> {
             case ConnectionState.active:
               if (snapshot.hasData) {
                 final allProfiles = snapshot.data as Iterable<CloudProfile>;
-                return ProfilesListView(
-                  profiles: allProfiles,
-                  onDeleteProfile: (profile) async {
-                    await _rentService.deleteProfile(id: profile.id);
-                  },
-                  onTap: (profile) {
-                    Navigator.of(context).push(
-                      MaterialPageRoute(
-                        builder: (context) => ReadProfile(profile: profile),
-                      ),
-                    );
-                  },
-                  onLongPress: (profile) {
-                    Navigator.of(context).push(
-                      MaterialPageRoute(
-                        builder: (context) =>
-                            CreateOrUpdateProfile(profile: profile),
-                      ),
-                    );
+                final profilesByCompanyId =
+                    _groupProfilesByCompanyId(allProfiles);
+
+                return FutureBuilder<Map<String, String>>(
+                  future: _getCompanyNames(profilesByCompanyId.keys.toList()),
+                  builder: (context, companyNamesSnapshot) {
+                    if (companyNamesSnapshot.connectionState ==
+                        ConnectionState.waiting) {
+                      return const Center(child: CircularProgressIndicator());
+                    }
+                    if (companyNamesSnapshot.hasData) {
+                      final companyNames = companyNamesSnapshot.data!;
+                      return ProfilesListView(
+                        groupedProfiles: profilesByCompanyId,
+                        companyNames: companyNames,
+                        onDeleteProfile: (profile) async {
+                          await _rentService.deleteProfile(id: profile.id);
+                        },
+                        onTap: (profile) {
+                          Navigator.of(context).push(
+                            MaterialPageRoute(
+                              builder: (context) =>
+                                  ReadProfile(profile: profile),
+                            ),
+                          );
+                        },
+                        onLongPress: (profile) {
+                          Navigator.of(context).push(
+                            MaterialPageRoute(
+                              builder: (context) =>
+                                  CreateOrUpdateProfile(profile: profile),
+                            ),
+                          );
+                        },
+                      );
+                    } else {
+                      return const Center(
+                          child: Text('Error loading company names'));
+                    }
                   },
                 );
               } else {
-                return const CircularProgressIndicator();
+                return const Center(child: Text('No profiles found'));
               }
             default:
-              return const CircularProgressIndicator();
+              return const Center(child: Text('Error loading profiles'));
           }
         },
       ),
     );
+  }
+
+  Future<Map<String, String>> _getCompanyNames(List<String> companyIds) async {
+    final companyNames = <String, String>{};
+    for (var companyId in companyIds) {
+      final company = await _rentService.getCompanyById(companyId: companyId);
+      companyNames[companyId] = company.companyName;
+    }
+    return companyNames;
+  }
+
+  Map<String, List<CloudProfile>> _groupProfilesByCompanyId(
+      Iterable<CloudProfile> profiles) {
+    final Map<String, List<CloudProfile>> groupedProfiles = {};
+    for (var profile in profiles) {
+      if (!groupedProfiles.containsKey(profile.companyId)) {
+        groupedProfiles[profile.companyId] = [];
+      }
+      groupedProfiles[profile.companyId]!.add(profile);
+    }
+    return groupedProfiles;
   }
 }
